prueba-master/requirements.txt
certifi==2022.5.18
cffi==1.15.1
chardet==4.0.0
cryptography==3.2.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.14
django-allauth==0.44.0
django-cors-headers==3.5.0
django-import-export==2.5.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
numpy==1.18.5
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==2.6.4
pandas==0.25.3
Pillow==7.2.0
psycopg2-binary==2.8.5
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tablib==2.0.0
urllib3==1.26.9
xlrd==1.2.0
xlwt==1.3.0
